Refactor navigation shared components
